Revert logic and exports separation to follow mod.
